refactor(plugin-sdk): sever protocol registry edges
