+++ src/infra/exec-approval-channel-runtime.ts
@@ -1,6 +1,6 @@
 // Runs the gateway-backed runtime that delivers native approval events.
 import { readConnectErrorDetailCode } from "../../packages/gateway-protocol/src/connect-error-details.js";
-import type { EventFrame } from "../../packages/gateway-protocol/src/index.js";
+import type { EventFrame } from "../../packages/gateway-protocol/src/schema/frames.js";
 import { startGatewayClientWhenEventLoopReady } from "../gateway/client-start-readiness.js";
 import type { GatewayClient, GatewayReconnectPausedInfo } from "../gateway/client.js";
 import { isApprovalMethod } from "../gateway/method-scopes.js";
